do a bit of initial cleanup
remove unused username field, change gender to use EnumConfigKey,
and move all the key classes to their own package
