+++ src/main/java/com/wildfire/main/config/AbstractConfiguration.java
@@ -24,6 +24,7 @@ import com.google.gson.JsonObject;
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonWriter;
 import com.wildfire.main.WildfireGender;
+import com.wildfire.main.config.types.ConfigKey;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.loader.api.FabricLoader;
 
@@ -67,13 +68,13 @@ public abstract class AbstractConfiguration {
 	}
 
 	public <TYPE> void setDefault(ConfigKey<TYPE> key) {
-		if(!SAVE_VALUES.has(key.key)) {
-			set(key, key.defaultValue);
+		if(!SAVE_VALUES.has(key.getKey())) {
+			set(key, key.getDefault());
 		}
 	}
 
 	public void removeParameter(ConfigKey<?> key) {
-		removeParameter(key.key);
+		removeParameter(key.getKey());
 	}
 
 	public void removeParameter(String key) {
