+++ src/main/java/com/wildfire/main/config/Configuration.java
@@ -18,17 +18,19 @@
 
 package com.wildfire.main.config;
 
-import java.nio.charset.StandardCharsets;
-import java.util.UUID;
+import com.wildfire.main.config.enums.Gender;
+import com.wildfire.main.config.types.BooleanConfigKey;
+import com.wildfire.main.config.types.EnumConfigKey;
+import com.wildfire.main.config.types.FloatConfigKey;
 
 public class Configuration extends AbstractConfiguration {
 
 	private static final String CONFIG_DIR = "WildfireGender";
 
-	public static final UUIDConfigKey USERNAME = new UUIDConfigKey("username", UUID.nameUUIDFromBytes("UNKNOWN".getBytes(StandardCharsets.UTF_8)));
-	public static final GenderConfigKey GENDER = new GenderConfigKey("gender");
+	public static final EnumConfigKey<Gender> GENDER = new EnumConfigKey<>("gender", Gender.MALE, Gender.BY_ID);
 	public static final FloatConfigKey BUST_SIZE = new FloatConfigKey("bust_size", 0.6F, 0, 0.8f);
 	public static final BooleanConfigKey HURT_SOUNDS = new BooleanConfigKey("hurt_sounds", true);
+	public static final FloatConfigKey VOICE_PITCH = new FloatConfigKey("voice_pitch", 1F, 0.8f, 1.2f);
 
 	public static final FloatConfigKey BREASTS_OFFSET_X = new FloatConfigKey("breasts_xOffset", 0.0F, -1, 1);
 	public static final FloatConfigKey BREASTS_OFFSET_Y = new FloatConfigKey("breasts_yOffset", 0.0F, -1, 1);
@@ -42,9 +44,6 @@ public class Configuration extends AbstractConfiguration {
 	public static final FloatConfigKey BOUNCE_MULTIPLIER = new FloatConfigKey("bounce_multiplier", 0.333F, 0, 0.5f);
 	public static final FloatConfigKey FLOPPY_MULTIPLIER = new FloatConfigKey("floppy_multiplier", 0.75F, 0.25f, 1);
 
-	public static final FloatConfigKey VOICE_PITCH = new FloatConfigKey("voice_pitch", 1F, 0.8f, 1.2f);
-
-	//Render holiday themes on the player.
 	public static final BooleanConfigKey HOLIDAY_THEMES = new BooleanConfigKey("holiday_themes", true);
 
 	public Configuration(String cfgName) {
