+++ src/main/java/com/wildfire/main/config/GlobalConfig.java
@@ -20,6 +20,9 @@ package com.wildfire.main.config;
 
 import com.wildfire.main.config.enums.ShowPlayerListMode;
 import com.wildfire.main.config.enums.SyncVerbosity;
+import com.wildfire.main.config.types.BooleanConfigKey;
+import com.wildfire.main.config.types.EnumConfigKey;
+import com.wildfire.main.config.types.StringConfigKey;
 
 public class GlobalConfig extends AbstractConfiguration {
     public static final GlobalConfig INSTANCE = new GlobalConfig();
