+++ src/main/java/com/wildfire/main/config/enums/Gender.java
@@ -16,8 +16,9 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.wildfire.main;
+package com.wildfire.main.config.enums;
 
+import com.wildfire.main.WildfireSounds;
 import io.netty.buffer.ByteBuf;
 import net.minecraft.network.codec.PacketCodec;
 import net.minecraft.network.codec.PacketCodecs;
