+++ src/main/java/com/wildfire/main/config/types/EnumConfigKey.java
@@ -16,31 +16,32 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.wildfire.main.config;
+package com.wildfire.main.config.types;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonPrimitive;
 
-public class StringConfigKey extends ConfigKey<String> {
+import java.util.function.IntFunction;
 
-    public StringConfigKey(String key, String defaultValue) {
-        super(key, defaultValue);
-    }
+public class EnumConfigKey<TYPE extends Enum<TYPE>> extends ConfigKey<TYPE> {
+	private final IntFunction<TYPE> ordinal;
 
-    @Override
-    protected String read(JsonElement element) {
-        if (element.isJsonPrimitive()) {
-            JsonPrimitive primitive = element.getAsJsonPrimitive();
-            if (primitive.isString()) {
-                return primitive.getAsString();
-            }
-        }
-        return defaultValue;
-    }
+	public EnumConfigKey(String key, TYPE defaultValue, IntFunction<TYPE> ordinalMapper) {
+		super(key, defaultValue);
+		this.ordinal = ordinalMapper;
+	}
 
-    @Override
-    public void save(JsonObject object, String value) {
-        object.addProperty(key, value);
-    }
+	@Override
+	protected TYPE read(JsonElement element) {
+		if(element instanceof JsonPrimitive prim && prim.isNumber()) {
+			return ordinal.apply(prim.getAsInt());
+		}
+		return defaultValue;
+	}
+
+	@Override
+	public void save(JsonObject object, TYPE value) {
+		object.addProperty(key, value.ordinal());
+	}
 }
