+++ src/main/java/com/wildfire/main/networking/ClientboundSyncPacket.java
@@ -18,7 +18,7 @@
 
 package com.wildfire.main.networking;
 
-import com.wildfire.main.Gender;
+import com.wildfire.main.config.enums.Gender;
 import com.wildfire.main.WildfireGender;
 import com.wildfire.main.entitydata.Breasts;
 import com.wildfire.main.entitydata.PlayerConfig;
